refactor: move app submission page under projects folder
