+++ src/containers/Projects/containers/AppSubmissionGuide/index.tsx
@@ -5,7 +5,7 @@ import AppSubmissionHero from './AppSubmissionHero';
 import AppSubmissionSteps from './AppSubmissionSteps';
 import AppSubmissionMessage from './AppSubmissionMessage';
 import AppSubmissionFaq from './AppSubmissionFaq';
-import Breadcrumb from '../Projects/components/Breadcrumb';
+import Breadcrumb from '../../components/Breadcrumb';
 
 const AppSubmissionGuide: FC = () => {
   return (
